Atualizando bibliotecas

+++ 2025/L_Ricci/src/lib/Magnetrometro/HMC5883L.h
@@ -9,8 +9,8 @@ private:
 public:
   bool begin();
   bool readAll();
-  float getX();
-  float getY();
-  float getZ();
+  int getX();
+  int getY();
+  int getZ();
 };
 #endif
